added list view

+++ app/src/main/java/com/gulij/moonslice/MainActivity.kt
@@ -4,6 +4,7 @@ import android.annotation.SuppressLint
 import android.content.Intent
 import android.graphics.drawable.Drawable
 import android.os.Bundle
+import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.gulij.moonslice.enums.Hemisphere
 import kotlinx.android.synthetic.main.activity_main.*
@@ -21,19 +22,28 @@ class MainActivity : AppCompatActivity() {
 
         Configuration.load(this)
 
-        val phase = MoonPhase.calculate(Configuration.instance.phaseAlgorithm, LocalDateTime.now()).toInt()
+        val phase = MoonPhase.calculate(Configuration.instance.phaseAlgorithm, LocalDateTime.now())
 
         todayTextView.text = "Dzisiaj: $phase%"
         moonImageView.setImageDrawable(getMoonDrawable(Configuration.instance.hemisphere, phase))
 
-        settingsButton.setOnClickListener { startActivity(Intent(this, SettingsActivity::class.java)) }
+        settingsButton.setOnClickListener {
+            startActivity(
+                Intent(
+                    this,
+                    SettingsActivity::class.java
+                )
+            )
+        }
     }
 
     private fun getMoonDrawable(hemisphere: Hemisphere, phase: Int): Drawable {
-        return getDrawable(resources.getIdentifier(
-            packageName + ":/drawable/"
-                    + (if (hemisphere == Hemisphere.north) "n" else "s") + phase.toString(),
-            null, null
-        ))!!
+        return getDrawable(
+            resources.getIdentifier(
+                packageName + ":/drawable/"
+                        + (if (hemisphere == Hemisphere.north) "n" else "s") + phase.toString(),
+                null, null
+            )
+        )!!
     }
 }
